% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnar_inject.R
\name{mnar_inject}
\alias{mnar_inject}
\title{Introduce NA's not at Random in a data.frame}
\usage{
mnar_inject(df, mnar_data)
}
\arguments{
\item{df}{data.frame, containing only the columns to be injected}

\item{mnar_data, }{MNARData object, containing the coefficients and predictors for the logit model
mtcars_mnar_data = MNARData(
             coef=matrix(c(2,-1,0, 1,0,2),3,2), 
             data=as.matrix(cbind(rep(1, nrow(mtcars)), mtcars[c('vs','am')])))
mtcars_na = mnar_inject(mtcars[c('hp','mpg')], mtcars_mnar_data)
head(mtcars_na)}
}
\description{
This function takes a data.frame and injects missing values not-at-random, but
according to a glm logit model. The \code{MNARData} object contains the coefficients and the predictor variables.
Note that the `data.frame` in the first argument should \emph{only} contain the variables that are to be injected.
This assumes no covariance between missingness for different variables.
Also note, this probably only works for binary categorical predictor variables.
}
