% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnardata.R
\docType{class}
\name{MNARData-class}
\alias{MNARData-class}
\alias{MNARData}
\title{Coefficients and Predictors for Introducing MNAR}
\arguments{
\item{coef}{a matrix of coefficients for a logit model for missingness}

\item{data}{predictor values for said logit model}
}
\description{
Coefficients and Predictors for Introducing MNAR
}
\examples{
# Only the first variable playes a role in missingness 
# Include the intercept column 
d <- MNARData(coef=matrix(c(2,-1,0), byrow=T), data=cbind(rep(1,5), replicate(2, rnorm(5))))
}
