% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_likert.R
\name{sim_likert}
\alias{sim_likert}
\title{Simulate a Likert data set with Missing Values and Misfitting Items}
\usage{
sim_likert(
  n_obs = 500,
  n_var = 5,
  low = 1,
  high = 5,
  n_misf = 2,
  perc_miss = 0.1,
  mnar = F,
  mnar_data,
  mnar_cols
)
}
\arguments{
\item{n_obs}{number of rows in the resultinf data.frame}

\item{n_var}{number of variables in the resulting data.frame (excluding misfiting items, see below}

\item{n_misf}{number of misfiting items}

\item{perc_miss}{percentage missing (for missing at-random)}

\item{mnar}{should missing not-at-random be injected (results in a separate DF)}

\item{mnar_data}{a MNARData object contains model definition for injecting MNARs}

\item{mnar_cols}{character vector of column names of the columns (obligatory)}

\item{lowest}{value for item response ?}

\item{highest}{value for item response ?}
}
\value{
A list with the requested data.frames
}
\description{
Misfitting items are for now loosely defined: They are just random responses without any structure.
A more sophisticated way will be to simulate items drawn from other structures (e.g. loading to more than one factors etc).
}
